Added Electronics Battery and Battery Box (See Desc)
Level of Detail change to LevelofDetail2 in primary assembly. there may be an issue with a contraint or model in the battery box assembly, as placing this assembly in the final submarine assembly causes severe system hangs and constraint corruption (may be a localized issue ie my machine, CPU usage is spiked heavily).